refactor: 컨트롤러를 준비, 실행, 결과 과정으로 분리한다
